// src/interfaces/petInterface.ts
export interface IPet{
    _id: string;
    avatar: string;
    name: string;   
    birthDate: string;
    breed: string;
    comments: string;
    owner: string;
}

export interface ICreatedPet {
    avatar: string;
    name: string; 
    birthDate: string;
    breed: string;
    comments: string;
    owner: string;
}